+++ MyApi/Controllers/TagController.cs
@@ -1,6 +1,7 @@
 using MyApi.Service;
 using Microsoft.AspNetCore.Mvc;
 using MyApi.Dto;
+using Microsoft.AspNetCore.Authorization;
 namespace MyApi.Controllers
 {
     [ApiController]
@@ -9,38 +10,73 @@ namespace MyApi.Controllers
     {
         private readonly ITagService _tagService;
 
-        public TagController(ITagService tagService)
+        private readonly IAuthorizationService _authorizationService;
+
+
+        public TagController(ITagService tagService, IAuthorizationService authorizationService)
         {
+            _authorizationService = authorizationService;
             _tagService = tagService;
         }
 
         [HttpGet]
+        [Authorize(Roles = "ADMIN")]
         public async Task<ActionResult<IEnumerable<TagDto>>> GetAllTags()
         {
             return Ok(await _tagService.GetAllTagsAsync());
         }
 
         [HttpGet("{id}")]
+        [Authorize]
         public async Task<ActionResult<TagDto>> GetTagById(int id)
         {
+            var tag = await _tagService.GetTagByIdAsync(id);
+            if (tag == null)
+            {
+                return NotFound();
+            }
+            var result = await _authorizationService.AuthorizeAsync(User, tag, "EditPolicy");
+            if (!result.Succeeded)
+            {
+                return Forbid();
+            }
             return Ok(await _tagService.GetTagByIdAsync(id));
         }
 
         [HttpPost]
+        [Authorize]
         public async Task<ActionResult<TagDto>> CreateTag([FromBody] TagDto tagDto)
         {
+            var result = await _authorizationService.AuthorizeAsync(User, tagDto, "EditPolicy");
+            if (!result.Succeeded)
+            {
+                return Forbid();
+            }
             return Ok(await _tagService.CreateTagAsync(tagDto));
         }
 
         [HttpPut("{id}")]
+        [Authorize]
         public async Task<ActionResult<TagDto>> UpdateTag([FromBody] TagDto tagDto)
         {
+            var result = await _authorizationService.AuthorizeAsync(User, tagDto, "EditPolicy");
+            if (!result.Succeeded)
+            {
+                return Forbid();
+            }
             return Ok(await _tagService.UpdateTagAsync(tagDto));
         }
 
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<ActionResult> DeleteTag(int id)
         {
+            var tag = await _tagService.GetTagByIdAsync(id);
+            var result = await _authorizationService.AuthorizeAsync(User, tag, "EditPolicy");
+            if (!result.Succeeded)
+            {
+                return Forbid();
+            }
             await _tagService.DeleteTagAsync(id);
             return NoContent();
         }
