+++ MyApi/Controllers/TodoController.cs
@@ -1,6 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 using MyApi.Dto;
 using MyApi.Service;
+using Microsoft.AspNetCore.Authorization;
 namespace MyApi.Controllers
 {
     [ApiController]
@@ -8,39 +9,76 @@ namespace MyApi.Controllers
     public class TodoController : ControllerBase
     {
         private readonly ITodoService _todoService;
+        private readonly IAuthorizationService _authorizationService;
 
-        public TodoController(ITodoService todoService)
+        public TodoController(ITodoService todoService, IAuthorizationService authorizationService)
         {
+            _authorizationService = authorizationService;
             _todoService = todoService;
         }
 
         [HttpGet]
+        [Authorize(Roles = "ADMIN")]
         public async Task<ActionResult<IEnumerable<TodoDto>>> GetAllTodos()
         {
             return Ok(await _todoService.GetAllTodosAsync());
         }
 
         [HttpGet("{id}")]
+        [Authorize]
         public async Task<ActionResult<TodoDto>> GetTodoById(int id)
         {
-            return Ok(await _todoService.GetTodoByIdAsync(id));
+            var todo = await _todoService.GetTodoByIdAsync(id);
+            if (todo == null)
+            {
+                return NotFound();
+            }
+            var result = await _authorizationService.AuthorizeAsync(User, todo, "EditPolicy");
+            if (!result.Succeeded)
+            {
+                return Forbid();
+            }
+            return Ok(todo);
         }
 
         [HttpPost]
+        [Authorize]
         public async Task<ActionResult<TodoDto>> CreateTodo([FromBody] TodoDto todoDto)
         {
+            var result = await _authorizationService.AuthorizeAsync(User, todoDto, "EditPolicy");
+            if (!result.Succeeded)
+            {
+                return Forbid();
+            }
             return Ok(await _todoService.CreateTodoAsync(todoDto));
         }
 
         [HttpPut("{id}")]
+        [Authorize]
         public async Task<ActionResult<TodoDto>> UpdateTodo([FromBody] TodoDto todoDto)
         {
+            var result = await _authorizationService.AuthorizeAsync(User, todoDto, "EditPolicy");
+            if (!result.Succeeded)
+            {
+                return Forbid();
+            }
             return Ok(await _todoService.UpdateTodoAsync(todoDto));
         }
 
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<ActionResult> DeleteTodoById(int id)
         {
+            var todo = await _todoService.GetTodoByIdAsync(id);
+            if (todo == null)
+            {
+                return NotFound();
+            }
+            var result = await _authorizationService.AuthorizeAsync(User, todo, "EditPolicy");
+            if (!result.Succeeded)
+            {
+                return Forbid();
+            }
             await _todoService.DeleteTodoByIdAsync(id);
             return NoContent();
         }
